ref: Move firebase file itno App subfolder
