+++ src/App/components/JsPsychExperiment.jsx
@@ -2,7 +2,7 @@ import { initJsPsych } from "jspsych";
 import React, { useEffect, useMemo, useRef } from "react";
 
 import { config } from "../../config/main";
-import { initParticipant } from "../../firebase";
+import { initParticipant } from "../deployments/firebase";
 import { buildTimeline, jsPsychOptions } from "../../timelines/main";
 
 export default function JsPsychExperiment({
