+++ src/index.js
@@ -5,4 +5,8 @@ import App from "./App/App.jsx";
 
 import "./App/index.css";
 
+/** Root of the application.
+ *
+ * This file renders the React application inside the given location (the browser or Electron)
+ */
 ReactDOM.render(<App />, document.getElementById("root"));
